requirements.txt
browser-use>=0.1.40
langchain-openai>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
cloudinary>=1.40.0
pillow>=10.0.0
